requirements.txt
Flask==0.10.1
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
wsgiref==0.1.2
gevent==1.0.1
tweepy==3.3.0
kafka-python
redis
uniout
flask-socketio
pymongo